Actualizacion tabla bicing_station y eliminacion de idioma en tabla user

+++ src/Entity/BicingStationEntity.cs
@@ -10,9 +10,9 @@ public class BicingStationEntity
 
     public required int BicingId { get; set; }
     public required string BicingName { get; set; }
-    public required float Latitude { get; set; }
-    public required float Longitude { get; set; }
-    public required float Altitude { get; set; }
+    public required double Latitude { get; set; }
+    public required double Longitude { get; set; }
+    public required double Altitude { get; set; }
     public required string Address { get; set; }
     public required string CrossStreet { get; set; }
     public required string PostCode { get; set; }
